Port the Merkle search tree core and wire to C++
Add a header-first C++20 implementation of the policy-bound B=16 Merkle search tree. The port provides strict canonical codecs, SHA-256 through CNG or OpenSSL, exact mst-sha256-b16-v2 domain framing, byte-identical MST2 blocks, persistent ordered-map updates, move-only representative handles, structural sharing diagnostics, semantic equality and diff, exact block inspection, and deep re-encoding validation.

Integrate an aggregate HAMT include, package-style copied-header consumer, separate Merkle test executable, and complete local API, usage, wire, validation, and test documentation. The native suite covers the shared golden vector, Unicode identifier rules, exact levels zero through four, independent histories, nullable values, equivalent representatives, a 12,000-operation retained model, callback exceptions, mutation detection, and concurrent readers.

Harden the implementation after independent cross-toolchain review by caching an entry level before moving the entry, making iterator frames comparable, decoding Unicode White_Space consistently, ordering framed-size overflow checks safely, handling empty OpenSSL input explicitly, and fixing the aggregate header's popcount widening warning.

Validation: MSVC Debug and Release both pass 36 legacy tests, 10 Merkle groups, and the copied-header consumer under /W4 /WX. GCC 16 and Clang 21 pass the same executables in O0 and O2 under -Wall -Wextra -Wpedantic -Werror; public headers and the consumer also pass -Wconversion -Wsign-conversion. Clang static analysis is silent, staged diff checks pass, and every C++ HAMT Markdown link resolves.

+++ src/Cpp/Hamt/include/Tools/DataStructures/Hamt/persistent_hash_map.hpp
@@ -677,7 +677,7 @@ private:
     }
 
     static std::size_t slot(std::uint32_t bitmap, std::uint32_t bit) noexcept {
-        return std::popcount(bitmap & (bit - 1u));
+        return static_cast<std::size_t>(std::popcount(bitmap & (bit - 1u)));
     }
 
     std::uint32_t get_hash(const Key& key) const {
